refactor(editor): share the source-link machinery between both windows
The Flow Graph grew the only code in the package that turns captured
diagnostic text back into something a reader can act on: the C# declaration
a message type was written in, the source line a stack frame points at, and
the button that opens either one. The Message Monitor needs all three for
issue #344, and copying roughly a thousand lines to get them would leave two
copies of a resolver whose retry, merge and backoff rules are subtle.

Moves that machinery out of `DxMessagingFlowGraphWindow` into
`DxMessagingEditorSourceLinks`: the lazy per-assembly source index and its
drain, the declaration parser, the stack-frame location parser, the emission
call-site reader, and the link button. `FlowGraphSourceLocation` becomes
`SourceLocation` and the link button's name loses its `flow-graph` segment,
because neither belongs to one window any more. Nothing is duplicated and no
forwarders are left behind -- the Flow Graph and its tests call the new home
directly.

Behavior-preserving by construction (a move, not a rewrite) and measured:
the focused Flow Graph fixture is 175/0 on the host editor after the move,
the same count it reported before it.
